// src/app/utils/types.ts
export interface CurrencyOption {
  value: string;
  id: number;
  type: string;
}

export type TokenCurrencyConversionRateMap = {
  [key in keyof typeof Token]?: TokenCurrencyConversionRate | undefined;
};

export enum Sign {
  POSITIVE,
  NEGATIVE,
  NEUTRAL,
}

export enum Currency {
  EUR = 'EUR',
  USD = 'USD',
}